+++ src/http/ui/src/index.tsx
@@ -4,13 +4,16 @@ import "./i18n";
 import App from "./App";
 import { AuthProvider } from "./context/AuthProvider";
 import { WebSocketProvider } from "./context/B4WsProvider";
+import { AiStatusProvider } from "./context/AiStatusProvider";
 
 const root = createRoot(document.getElementById("root")!);
 root.render(
   <BrowserRouter>
     <AuthProvider>
       <WebSocketProvider>
-        <App />
+        <AiStatusProvider>
+          <App />
+        </AiStatusProvider>
       </WebSocketProvider>
     </AuthProvider>
   </BrowserRouter>,